refactor(desktop): settings-view 收口为路由壳并完成目录整理
将 SettingsView 迁至 settings/settings-view.tsx，根目录 settings-view.tsx 改为 re-export 桶。
